implement actor registration in Runtime
The runtime now exposes a `create_and_register()` function for
instantiating components as Actors and registering their references with
the dispatcher.

This registration happens by enqueing a registration message on the
dispatcher's message queue, allowing the dispatcher to handle the
message as it sees fit. Deregistration happens the same way.